static_analyzer: move functions to oswactcher utils
