+++ oswatcher/utils/__init__.py
@@ -8,3 +8,17 @@ def get_hard_drive_path(domain):
         raise RuntimeError('Cannot find hard disk for domain {}'.format(domain.name()))
     qcow_path = disk.find('source').get('file')
     return qcow_path
+
+
+def format_size(self, size: int, precision: int = 2) -> str:
+    suffix = ['B', 'KB', 'MB', 'GB']
+    suffix_index = 0
+
+    if size == 0:
+        return "0"
+    else:
+        while size > 1024 and suffix_index < 3:
+            suffix_index += 1
+            size = size / 1024.0
+
+    return "%.*f%s" % (precision, size, suffix[suffix_index])
